moved EMAIL_AGENT to types
