--- src/components/notifications/constants.ts
@@ -1,4 +1,0 @@
-export const EMAIL_AGENT = {
-    SES: 'SES',
-    SMTP: 'SMTP',
-}
